+++ src/components/project/TileView.svelte
@@ -544,7 +544,7 @@
         display: flex;
         flex-direction: row;
         flex-wrap: nowrap;
-        overflow: auto;
+        overflow: hidden;
     }
 
     .main.rtl {