+++ CHANGELOG.md
@@ -20,6 +20,7 @@ These notes are publicly posted in [production](https://wordplay.dev/updates), s
 ### Fixed
 
 - We found and fixed a problem where the autocomplete menu wouldn't appear when activated in the menu.
+- We removed some unnecessary scroll bars (#1098).
 
 ## 0.17.8 - 2026-05-02